Debug

+++ GameEntity/Bullet.h
@@ -10,7 +10,7 @@ class Bullet : public QObject, public QGraphicsRectItem, public GameEntity
 {
     Q_OBJECT
 public:        
-    Bullet(Tank* tank, const double& damage, const double& degree, const int& size, const int& vx, const int& vy );
+    Bullet(Tank* tank, const double& damage, const double& degree, const int& size, const double& vx, const double& vy );
     double get_damage() const;
     double get_degree() const {return degree;}
     void set_degree (double degree) {this->degree = degree;}
